RPi4: Resolve print formatting errors

+++ components/StorageInterfaceTester/test_storage.c
@@ -4,6 +4,7 @@
 #include "TestMacros.h"
 
 #include <stdlib.h>
+#include <sys/types.h>
 
 static const off_t storageBeginOffset = 0U;
 
@@ -113,7 +114,7 @@ roundDownToBLockSize(
     if(OS_ERROR_NOT_IMPLEMENTED == rslt) \
     { \
         /* Erase functionality is considered optional. */ \
-        ASSERT_EQ_INT_MAX(0LL, bytesErased); \
+        ASSERT_EQ_INT_MAX(0L, bytesErased); \
 \
         Debug_LOG_WARNING( \
             "Erase function is not implemented for %s. Was it intended?", \
@@ -183,7 +184,7 @@ test_storage_blockSize_pos()
 
     // Different storages have different block sizes, but nevertheless the block
     // size must always be greater than 0.
-    ASSERT_LT_SZ(0U, storageBlockSize);
+    ASSERT_LT_SZ((size_t)0U, storageBlockSize);
 
     TEST_FINISH();
 }
@@ -243,8 +244,8 @@ test_storage_writeReadEraseZeroBytes_pos()
 
     const off_t testOffset = 0;
 
-    TEST_WRITE(testOffset, testData, 0U);
-    TEST_READ (testOffset, testData, 0U);
+    TEST_WRITE(testOffset, testData, (size_t)0U);
+    TEST_READ (testOffset, testData, (size_t)0U);
 
     off_t bytesErased = -1;
 
@@ -260,7 +261,7 @@ test_storage_writeReadEraseZeroBytes_pos()
         TEST_SUCCESS(rslt);
     }
 
-    ASSERT_EQ_INT_MAX(0LL, bytesErased);
+    ASSERT_EQ_INT_MAX(0L, bytesErased);
 
     TEST_FINISH();
 }
@@ -389,7 +390,7 @@ test_storage_neighborRegionsUntouched_pos()
     TEST_NOT_SUCCESS( \
         storage_rpc_write(offset, roundedDownSize, &bytesWritten)); \
 \
-    ASSERT_EQ_SZ(0U, bytesWritten); \
+    ASSERT_EQ_SZ((size_t)0U, bytesWritten); \
 } while (0)
 
 #define TEST_READ_NEG(offset, size) do \
@@ -407,7 +408,7 @@ test_storage_neighborRegionsUntouched_pos()
     TEST_NOT_SUCCESS( \
         storage_rpc_read(offset, roundedDownSize, &bytesRead)); \
 \
-    ASSERT_EQ_SZ(0U, bytesRead); \
+    ASSERT_EQ_SZ((size_t)0U, bytesRead); \
 } while (0)
 
 #define TEST_ERASE_NEG(offset, size) do \
@@ -425,7 +426,7 @@ test_storage_neighborRegionsUntouched_pos()
             roundDownToBLockSize(size), \
             &bytesErased)); \
 \
-    ASSERT_EQ_INT_MAX(0LL, bytesErased); \
+    ASSERT_EQ_INT_MAX(0L, bytesErased); \
 } while (0)
 
 #define TEST_WRITE_READ_ERASE_NEG(offset, size) do \
